Update Request java samples
Pretty confident for most of these. The PUT and the POST (change and
create update request) need some additional code that I am not sure
about. This is also the doc that had the pending changes, i.e., create
update request. But those pending changes are so old now that the
method has been deprecated.