Edit comments works in C#
Took out the Java example I made as I think there needs to be an update
to the SDK. It didn’t work with the AddCommentBuilder and didn’t work
when I removed the Builder. I can only get it to send a POST instead of
the PUT that it needs.