Removed link to GitHub Getting Started
Page is going away. Background: this was an interim rewrite of Getting Started before the major overhaul in summer of 2017. At this point, the doc is no longer needed. Some of the FAQ will move to the Dev Portal. Other content is covered in the API Docs in various places.